third_party/WebKit/LayoutTests/fast/events/wheel/wheelevent-in-scrolling-div-expected.txt

PASS event.deltaY is window.expectedScrollTop
PASS event.deltaX is window.expectedScrollLeft
PASS event.wheelDeltaY is -5*120
PASS event.wheelDeltaX is -2.5*120
PASS event.wheelDelta is -5*120
PASS div.scrollTop == window.expectedScrollTop && div.scrollLeft == window.expectedScrollLeft became true
PASS successfullyParsed is true

TEST COMPLETE
